Fix codesmell

+++ src/AveragingInterpolation.cpp
@@ -321,7 +321,7 @@ void meshkernel::AveragingInterpolation::ComputeOnPolygon(const std::vector<Poin
 
     if (m_method == Method::SimpleAveraging && numValidSamplesInPolygon > 0 && result > doubleMissingValue)
     {
-        result /= numValidSamplesInPolygon;
+        result /= static_cast<size_t>(numValidSamplesInPolygon);
     }
 
     if (m_method == Method::InverseWeightedDistance && numValidSamplesInPolygon > 0)
